move sync types to sync module breaking circular dep
