+++ src/db/mod.rs
@@ -11,8 +11,8 @@ pub mod test_helpers;
 pub use root::{get_root, GetRootError};
 
 pub use commit::{
-    BaseSnapshotError, Commit, FromHashError, LocalMeta, MetaTyped, PendingError, InternalProgrammerError,
-    DEFAULT_HEAD_NAME,
+    BaseSnapshotError, Commit, FromHashError, InternalProgrammerError, LocalMeta, MetaTyped,
+    PendingError, DEFAULT_HEAD_NAME,
 };
 pub use read::{read_commit, OwnedRead, Read, ReadCommitError, Whence};
 pub use scan::{ScanBound, ScanKey, ScanOptions};
